ml-service/api/routers/planogram.py
# ml-service/api/routers/planogram.py
# Planogram compliance analysis endpoints

from fastapi import APIRouter, File, UploadFile, HTTPException
from typing import Dict, Any
from PIL import Image
import io

router = APIRouter()

@router.post("/analyze")
async def analyze_planogram_compliance(
    shelf_image: UploadFile = File(...),
    planogram_image: UploadFile = File(...)
) -> Dict[str, Any]:
    """
    Analyze shelf compliance with planogram
    
    Args:
        shelf_image: Current shelf state image
        planogram_image: Expected planogram layout
    
    Returns:
        Compliance score and violations
    """
    try:
        # Read images
        shelf_data = await shelf_image.read()
        planogram_data = await planogram_image.read()
        
        shelf_img = Image.open(io.BytesIO(shelf_data))
        planogram_img = Image.open(io.BytesIO(planogram_data))
        
        # TODO: Implement planogram compliance analysis
        result = {
            "compliance_score": 0.87,  # 0-1
            "violations": [
                {
                    "type": "misplaced_product",
                    "location": [0.5, 0.3, 0.1, 0.15],
                    "severity": "medium",
                    "description": "Product in wrong position"
                }
            ],
            "total_violations": 1,
            "recommendations": [
                "Move Coca-Cola to position A3"
            ]
        }
        
        return result
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Planogram analysis failed: {str(e)}")

@router.post("/compare")
async def compare_shelf_states(
    before_image: UploadFile = File(...),
    after_image: UploadFile = File(...)
) -> Dict[str, Any]:
    """
    Compare two shelf states to detect changes
    """
    try:
        # Read images
        before_data = await before_image.read()
        after_data = await after_image.read()
        
        before_img = Image.open(io.BytesIO(before_data))
        after_img = Image.open(io.BytesIO(after_data))
        
        # TODO: Implement shelf state comparison
        result = {
            "changes_detected": True,
            "changes": [
                {
                    "type": "removed",
                    "location": [0.2, 0.3, 0.1, 0.15],
                    "description": "Product removed"
                }
            ],
            "similarity_score": 0.75
        }
        
        return result
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Comparison failed: {str(e)}")
