ml-service/requirements.txt
# ShelfSense ML Service
# Python FastAPI microservice for computer vision and ML inference

# Core ML/AI libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8/v11
transformers>=4.35.0  # For CLIP and other models
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0  # Extra OpenCV modules
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0

# CLIP for product classification
git+https://github.com/openai/CLIP.git
ftfy>=6.1.0
regex>=2023.10.0

# API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Image processing & augmentation
albumentations>=1.3.0  # Data augmentation
scikit-image>=0.21.0
imgaug>=0.4.0

# Vector database for embeddings
faiss-cpu>=1.7.0  # Use faiss-gpu for GPU support
chromadb>=0.4.0

# Database & Storage
psycopg2-binary>=2.9.0  # PostgreSQL
redis>=5.0.0
boto3>=1.28.0  # AWS S3 for image storage
motor>=3.3.0  # Async MongoDB

# HTTP & WebSockets
httpx>=0.25.0
websockets>=12.0
aiohttp>=3.9.0
python-socketio>=5.10.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.66.0
pyyaml>=6.0.0

# Monitoring & Metrics
prometheus-client>=0.19.0
psutil>=5.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0  # JWT tokens
passlib[bcrypt]>=1.7.4
httpx>=0.25.0  # HTTP client
aiofiles>=23.0.0  # Async file operations

# Monitoring & Logging
prometheus-client>=0.18.0
python-json-logger>=2.0.0

# Model optimization (optional)
onnx>=1.15.0
onnxruntime>=1.16.0
